Move ServiceHubRemoteHostClient down to EditorFeatures
