+++ src/VisualStudio/Core/Def/Microsoft.VisualStudio.LanguageServices.csproj
@@ -45,18 +45,6 @@
     <Compile Include="..\..\..\Compilers\Core\Portable\InternalUtilities\ConcurrentLruCache.cs">
       <Link>Shared\ConcurrentLruCache.cs</Link>
     </Compile>
-    <Compile Include="..\..\..\Workspaces\Remote\ServiceHub\Shared\RemoteEndPoint.cs">
-      <Link>Shared\RemoteEndPoint.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\Workspaces\Remote\ServiceHub\Shared\RoslynJsonConverter.cs">
-      <Link>Shared\RoslynJsonConverter.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\Workspaces\Remote\ServiceHub\Shared\RoslynJsonConverter.RoslynOnly.cs">
-      <Link>Shared\RoslynJsonConverter.RoslynOnly.cs</Link>
-    </Compile>
-    <Compile Include="..\..\..\Workspaces\Remote\ServiceHub\Shared\RoslynJsonConverter.SolutionIdConverters.cs">
-      <Link>Shared\RoslynJsonConverter.SolutionIdConverters.cs</Link>
-    </Compile>
     <Compile Include="..\..\..\ExpressionEvaluator\Core\Source\ExpressionCompiler\DkmExceptionUtilities.cs" Link="Implementation\EditAndContinue\Interop\DkmExceptionUtilities.cs" />
     <Compile Update="VSPackage.Designer.cs">
       <AutoGen>True</AutoGen>
