--- src/VisualStudio/Core/Def/Telemetry/RoslynTelemetrySetup.cs
@@ -1,31 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Diagnostics;
-using Microsoft.CodeAnalysis.Internal.Log;
-using Microsoft.CodeAnalysis.Options;
-using Microsoft.VisualStudio.ComponentModelHost;
-using Microsoft.VisualStudio.Telemetry;
-
-namespace Microsoft.VisualStudio.LanguageServices.Telemetry
-{
-    internal static class RoslynTelemetrySetup
-    {
-        public static void Initialize(IServiceProvider serviceProvider, TelemetrySession session)
-        {
-            var componentModel = (IComponentModel)serviceProvider.GetService(typeof(SComponentModel));
-            var optionService = componentModel.GetService<IGlobalOptionService>();
-
-            Logger.SetLogger(
-                AggregateLogger.Create(
-                    CodeMarkerLogger.Instance,
-                    new EtwLogger(optionService),
-                    new VSTelemetryLogger(session),
-                    Logger.GetLogger()));
-
-            Logger.Log(FunctionId.Run_Environment, KeyValueLogMessage.Create(m => m["Version"] = FileVersionInfo.GetVersionInfo(typeof(VisualStudioWorkspace).Assembly.Location).FileVersion));
-        }
-    }
-}
