+++ src/Workspaces/Core/Portable/Remote/IRemoteHostService.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+#nullable enable
+
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
@@ -11,7 +13,9 @@ namespace Microsoft.CodeAnalysis.Remote
 {
     internal interface IRemoteHostService
     {
-        void InitializeGlobalState(string host, int uiCultureLCID, int cultureLCID, string serializedSession, CancellationToken cancellationToken);
+        void InitializeGlobalState(int uiCultureLCID, int cultureLCID, CancellationToken cancellationToken);
+
+        void InitializeTelemetrySession(string host, string serializedSession, CancellationToken cancellationToken);
 
         /// <summary>
         /// This is only for debugging
